pyproject.toml
[tool.poetry]
name = "formula"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1"
django-celery-beat = "^2.7"
django-debug-toolbar = "^5.0"
django-guardian = "^2.4"
django-import-export = "^4.3"
django-simple-history = "^3.8"
django-modeltranslation = "^0.19"
django-money = "^3.5"
django-unfold = { git = "https://github.com/unfoldadmin/django-unfold.git" }
whitenoise = "^6.8"
gunicorn = "^23.0"
pillow = "^11.1"
sentry-sdk = { extras = ["django"], version = "^2.20" }
pygments = "^2.19"
python-dotenv = "^1.0.1"

[tool.ruff]
fix = true
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
]
exclude = ["**/migrations"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
